Validate dynamic linking against spdlog and fmt (#29)
While rapids-logger packages need to use specific approaches to building
and linking spdlog and fmt to avoid exposing undesirable symbols
publicly, a direct consumer of rapids-logger should be able to
dynamically link to preexisting spdlog and fmt libraries if they wish.
This PR ensures that is possible by adding suitable tests.

This PR also bumps the minimum required CMake version to match the rest
of RAPIDS.

Resolves rapidsai/cudf#16951