Fix fmt fetching (#51)
I had previously assumed we would be able to release the unconditional
switch over to std::format along with the migration away from
rapids-cmake's fmt and spdlog modules, so I didn't test this interaction
very thoroughly. Now that we are making a 0.2.x line without that, we
need to fix bugs in that interop.